auhentication added, only logged in authors can post

+++ routes/web.php
@@ -14,29 +14,31 @@ use App\Http\Controllers\HomeController;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
-Route::get('/', function () {
-    return view('welcome');
-});
+// Home page: show all paintings
+Route::get('/', [ListingController::class, 'index']);
 
-
-Route::get('/',[ListingController::class, 'index']);
+// Search
 Route::post('/search', [ListingController::class, 'search']);
-Route::get('/create',[ListingController::class, 'create']);
-Route::post('/gallery', [ListingController::class, 'store']);
 
-Route::delete('/gallery/{id}', [ListingController::class, 'destroy'])->name('gallery.destroy');
-Route::get('/gallery/{listing}/edit', [ListingController::class, 'edit']);
-Route::put('/gallery/{listing}', [ListingController::class, 'update']);
-Route::delete('/gallery/{listing}', [ListingController::class, 'destroy']);
+// Create post (only for authenticated users)
+Route::get('/create', [ListingController::class, 'create'])->middleware('auth');
+Route::post('/gallery', [ListingController::class, 'store'])->middleware('auth');
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
-Route::get('/gallery/{listing}',[ListingController::class, 'show']);
+// Edit, update, delete gallery items
+Route::get('/gallery/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('/gallery/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('/gallery/{listing}', [ListingController::class, 'destroy'])->name('gallery.destroy')->middleware('auth');
+
+// Show single gallery item
+Route::get('/gallery/{listing}', [ListingController::class, 'show']);
+
+// Home page for authenticated users (if needed)
 Route::get('/home', [HomeController::class, 'index']);
-Route::post('/users', [UserController::class, 'store']);
 
-Route::get('/login', [UserController::class, 'login'])->name('login');
-Route::post('/login', [UserController::class, 'authenticate'])->name('login.authenticate');
-
+// User registration & authentication
 Route::get('/register', [UserController::class, 'create'])->name('register');
 Route::post('/register', [UserController::class, 'store']);
+Route::get('/login', [UserController::class, 'login'])->name('login');
+Route::post('/login', [UserController::class, 'authenticate'])->name('login.authenticate');
 Route::post('/logout', [UserController::class, 'logout'])->name('logout');
-
